replace default implementation
